replace any in distsql into MetaOnlyInfoSchema
