+++ pkg/distsql/request_builder.go
@@ -27,7 +27,7 @@ import (
 	"github.com/pingcap/tidb/pkg/ddl/placement"
 	distsqlctx "github.com/pingcap/tidb/pkg/distsql/context"
 	"github.com/pingcap/tidb/pkg/errctx"
-	"github.com/pingcap/tidb/pkg/infoschema"
+	infoschema "github.com/pingcap/tidb/pkg/infoschema/context"
 	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/parser/mysql"
@@ -45,7 +45,7 @@ import (
 // It is called before we issue a kv request by "Select".
 type RequestBuilder struct {
 	kv.Request
-	is  infoschema.InfoSchema
+	is  infoschema.MetaOnlyInfoSchema
 	err error
 }
 
@@ -340,11 +340,7 @@ func (builder *RequestBuilder) SetTiDBServerID(serverID uint64) *RequestBuilder 
 
 // SetFromInfoSchema sets the following fields from infoSchema:
 // "bundles"
-func (builder *RequestBuilder) SetFromInfoSchema(pis any) *RequestBuilder {
-	is, ok := pis.(infoschema.InfoSchema)
-	if !ok {
-		return builder
-	}
+func (builder *RequestBuilder) SetFromInfoSchema(is infoschema.MetaOnlyInfoSchema) *RequestBuilder {
 	builder.is = is
 	builder.Request.SchemaVar = is.SchemaMetaVersion()
 	return builder
@@ -387,13 +383,13 @@ func (builder *RequestBuilder) verifyTxnScope() error {
 		if !valid {
 			var tblName string
 			var partName string
-			tblInfo, _, partInfo := builder.is.FindTableByPartitionID(phyTableID)
+			tblInfo, _, partInfo := builder.is.FindTableInfoByPartitionID(phyTableID)
 			if tblInfo != nil && partInfo != nil {
-				tblName = tblInfo.Meta().Name.String()
+				tblName = tblInfo.Name.String()
 				partName = partInfo.Name.String()
 			} else {
-				tblInfo, _ = builder.is.TableByID(phyTableID)
-				tblName = tblInfo.Meta().Name.String()
+				tblInfo, _ = builder.is.TableInfoByID(phyTableID)
+				tblName = tblInfo.Name.String()
 			}
 			err := fmt.Errorf("table %v can not be read by %v txn_scope", tblName, txnScope)
 			if len(partName) > 0 {
@@ -703,7 +699,7 @@ func CommonHandleRangesToKVRanges(dctx *distsqlctx.DistSQLContext, tids []int64,
 }
 
 // VerifyTxnScope verify whether the txnScope and visited physical table break the leader rule's dcLocation.
-func VerifyTxnScope(txnScope string, physicalTableID int64, is infoschema.InfoSchema) bool {
+func VerifyTxnScope(txnScope string, physicalTableID int64, is infoschema.MetaOnlyInfoSchema) bool {
 	if txnScope == "" || txnScope == kv.GlobalTxnScope {
 		return true
 	}
