+++ pkg/infoschema/context/BUILD.bazel
@@ -5,5 +5,8 @@ go_library(
     srcs = ["infoschema.go"],
     importpath = "github.com/pingcap/tidb/pkg/infoschema/context",
     visibility = ["//visibility:public"],
-    deps = ["//pkg/parser/model"],
+    deps = [
+        "//pkg/ddl/placement",
+        "//pkg/parser/model",
+    ],
 )
